Edit Vehicle now works

+++ fuel-monitor/src/org/feup/fuelmonitor/AddVehicle.java
@@ -6,6 +6,7 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.util.Calendar;
 
 import android.content.Intent;
 import android.database.Cursor;
@@ -34,6 +35,7 @@ public class AddVehicle extends SherlockActivity {
 	private FuelMonitorDbAdapter mDbHelper;
 	private static File mTempFile;
 	private boolean edit;
+	private long mVehicleID;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
@@ -46,7 +48,7 @@ public class AddVehicle extends SherlockActivity {
 		final TextView model = (TextView) findViewById(R.id.addvehicle_modelText);
 		final Spinner fuelType = (Spinner) findViewById(R.id.addvehicle_fuelTypeSpinner);
 		final TextView capacity = (TextView) findViewById(R.id.addvehicle_fuelCapacityText);
-		final TextView license = (TextView) findViewById(R.id.addvehicle_registrationText);
+		final TextView registration = (TextView) findViewById(R.id.addvehicle_registrationText);
 		final Spinner year = (Spinner) findViewById(R.id.addvehicle_yearSpinner);
 		final TextView kms = (TextView) findViewById(R.id.addvehicle_kmsText);
 		final CheckBox photo = (CheckBox) findViewById(R.id.addvehicle_photoCheckBox);
@@ -54,10 +56,30 @@ public class AddVehicle extends SherlockActivity {
 
 		mDbHelper.open();
 
+		if (edit) {
+			mVehicleID = getIntent().getLongExtra("vehicleID", 0);
+			Cursor editVehicle = mDbHelper.getVehicleByID(mVehicleID);
+			editVehicle.moveToFirst();
+			make.setSelection(editVehicle.getInt(editVehicle
+					.getColumnIndex("idMake")) + 1);
+			model.setText(editVehicle.getString(editVehicle
+					.getColumnIndex("model")));
+			fuelType.setSelection(editVehicle.getInt(editVehicle
+					.getColumnIndex("idFuelType")) + 1);
+			capacity.setText(Integer.toString(editVehicle.getInt(editVehicle
+					.getColumnIndex("fuelCapacity"))));
+			registration.setText(editVehicle.getString(editVehicle
+					.getColumnIndex("registration")));
+			year.setSelection(Calendar.getInstance().get(Calendar.YEAR)
+					- editVehicle.getInt(editVehicle.getColumnIndex("year")));
+			kms.setText(Integer.toString(editVehicle.getInt(editVehicle
+					.getColumnIndex("kms"))));
+		}
+
 		ArrayAdapter<Integer> yearAdapter = new ArrayAdapter<Integer>(this,
 				android.R.layout.simple_spinner_item);
 
-		for (int y = 2012; y >= 1900; y--)
+		for (int y = Calendar.getInstance().get(Calendar.YEAR); y >= 1900; y--)
 			yearAdapter.add(y);
 
 		Cursor makeCursor = mDbHelper.fetchMakes();
@@ -84,13 +106,64 @@ public class AddVehicle extends SherlockActivity {
 
 		save.setOnClickListener(new View.OnClickListener() {
 			public void onClick(View v) {
-				if (edit) {
-				} else {
-					if (!model.getText().toString().equals("")
-							&& !license.getText().toString().equals("")
-							&& !capacity.getText().toString().equals("")
-							&& !kms.getText().toString().equals("")) {
-						if (mDbHelper.addVehicle(
+				if (!model.getText().toString().equals("")
+						&& !registration.getText().toString().equals("")
+						&& !capacity.getText().toString().equals("")
+						&& !kms.getText().toString().equals("")) {
+					long queryRetCode;
+					if (edit) {
+						String oldReg = mDbHelper
+								.getRegistrationByID(mVehicleID);
+						queryRetCode = mDbHelper
+								.editVehicle(
+										mVehicleID,
+										((SimpleCursorAdapter) make
+												.getAdapter())
+												.getCursor()
+												.getLong(
+														((SimpleCursorAdapter) make
+																.getAdapter())
+																.getCursor()
+																.getColumnIndex(
+																		"_id")),
+										model.getText().toString(),
+										((SimpleCursorAdapter) fuelType
+												.getAdapter())
+												.getCursor()
+												.getLong(
+														((SimpleCursorAdapter) fuelType
+																.getAdapter())
+																.getCursor()
+																.getColumnIndex(
+																		"_id")),
+										Short.parseShort(capacity.getText()
+												.toString()), registration
+												.getText().toString(), Short
+												.parseShort(year
+														.getSelectedItem()
+														.toString()), Integer
+												.parseInt(kms.getText()
+														.toString()));
+						// change photo file names
+						if (queryRetCode > 0) {
+							File directory = new File(Environment
+									.getExternalStorageDirectory(),
+									"fuelmonitor/");
+							String newReg = mDbHelper
+									.getRegistrationByID(mVehicleID);
+							File oldimgFile = new File(directory,
+									(oldReg + ".jpg"));
+							File oldthumbFile = new File(directory,
+									(oldReg + "t.jpg"));
+							File newimgFile = new File(directory,
+									(newReg + ".jpg"));
+							File newthumbFile = new File(directory,
+									(newReg + "t.jpg"));
+							oldimgFile.renameTo(newimgFile);
+							oldthumbFile.renameTo(newthumbFile);
+						}
+					} else {
+						queryRetCode = mDbHelper.addVehicle(
 								((SimpleCursorAdapter) make.getAdapter())
 										.getCursor().getLong(
 												((SimpleCursorAdapter) make
@@ -105,51 +178,52 @@ public class AddVehicle extends SherlockActivity {
 														.getCursor()
 														.getColumnIndex("_id")),
 								Short.parseShort(capacity.getText().toString()),
-								license.getText().toString(), Short
+								registration.getText().toString(), Short
 										.parseShort(year.getSelectedItem()
 												.toString()), Integer
-										.parseInt(kms.getText().toString())) > 0) {
-							if (photo.isChecked()) {
-								File directory = new File(Environment
-										.getExternalStorageDirectory(),
-										"fuelmonitor/");
-								directory.mkdirs();
-								File file = new File(directory, (license
-										.getText().toString() + ".jpg"));
-								file.delete();
-								try {
-									file.createNewFile();
-								} catch (IOException e) {
-									Log.e(TAG, "Error creating image file");
-								}
-								Intent i = new Intent(
-										android.provider.MediaStore.ACTION_IMAGE_CAPTURE);
-								i.putExtra(MediaStore.EXTRA_OUTPUT,
-										Uri.fromFile(file));
-								mTempFile = new File(directory, (license
-										.getText().toString()));
-								startActivityForResult(i, 1);
-							} else
-								finish();
-						} else {
-							int duration = Toast.LENGTH_SHORT;
-
-							Toast toast = Toast
-									.makeText(
-											getApplicationContext(),
-											getString(R.string.add_vehicle_error_inserting_toast),
-											duration);
-							toast.show();
-						}
-
+										.parseInt(kms.getText().toString()));
+					}
+					if (queryRetCode > 0) {
+						if (photo.isChecked()) {
+							File directory = new File(Environment
+									.getExternalStorageDirectory(),
+									"fuelmonitor/");
+							directory.mkdirs();
+							File file = new File(directory, (registration
+									.getText().toString() + ".jpg"));
+							file.delete();
+							try {
+								file.createNewFile();
+							} catch (IOException e) {
+								Log.e(TAG, "Error creating image file");
+							}
+							Intent i = new Intent(
+									android.provider.MediaStore.ACTION_IMAGE_CAPTURE);
+							i.putExtra(MediaStore.EXTRA_OUTPUT,
+									Uri.fromFile(file));
+							mTempFile = new File(directory, (registration
+									.getText().toString()));
+							startActivityForResult(i, 1);
+						} else
+							finish();
 					} else {
 						int duration = Toast.LENGTH_SHORT;
 
-						Toast toast = Toast.makeText(getApplicationContext(),
-								getString(R.string.add_vehicle_fill_all_toast),
-								duration);
+						Toast toast = Toast
+								.makeText(
+										getApplicationContext(),
+										getString(R.string.add_vehicle_error_inserting_toast),
+										duration);
 						toast.show();
 					}
+
+				} else {
+					int duration = Toast.LENGTH_SHORT;
+
+					Toast toast = Toast.makeText(getApplicationContext(),
+							getString(R.string.add_vehicle_fill_all_toast),
+							duration);
+					toast.show();
 				}
 			}
 		});
